fix exception correction

+++ homework_6/src/Image.cpp
@@ -13,10 +13,10 @@ Image::Image() = default;
 Image::Image(int rows, int cols) : rows_{rows}, cols_{cols} {}
 int Image::cols() const { return cols_; }
 int Image::rows() const { return rows_; }
-int Image::at(int row, int col) const {
+uint8_t Image::at(int row, int col) const {
   return data_.at(this->rows_ * row + col);
 }
-int& Image::at(int row, int col) {
+uint8_t& Image::at(int row, int col) {
   return this->data_.at(this->rows_ * row + col);
 }
